Corrige build.gradle.kts para compatibilidade do pipeline

+++ android/app/build.gradle.kts
@@ -26,8 +26,9 @@ android {
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = flutter.minSdkVersion
         targetSdk = flutter.targetSdkVersion
-        versionCode = flutter.versionCode
+        versionCode = flutter.versionCode.toInt()
         versionName = flutter.versionName
+
     }
 
     buildTypes {
